remove deprecated use of `URI.encode`
`URI.encode` is removed in ruby 3.0; this makes the code work without
it.

Generally speaking, this change won't produce exactly identical query
strings, (although I think it will for "simple" `options` that are most
likely to be used, and that where the result is different the new result
will be more correct).

For example, given the options mentioned in the documentation comment:
  `options = { test: true, download: false, flatten: 1 }`

both the old code and the new code will produce
  `"?test=true&download=false&flatten=1"`

But with more "complex" options:
  `options = { search: "Tom & Jerry", page: 3 }`,

the old code will produce
  `"?search=Tom%20&%20Jerry&page=3"`

and the new code will produce
  `"?search=Tom+%26+Jerry&page=3"`